requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
pyreadr==0.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.0.3
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
pyreadr==0.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.0.3
alabaster==0.7.12
anndata==0.10.7
array_api_compat==1.6
Babel==2.10.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudpickle==3.0.0
cna==0.1.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.4
docutils==0.17.1
filelock==3.7.1
flake8==4.0.1
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.9
GitPython==3.1.27
h5py==3.11.0
harmony-pytorch==0.1.8
idna==3.3
imagesize==1.3.0
Jinja2==3.1.2
joblib==1.4.0
kiwisolver==1.4.5
legacy-api-wrap==1.4
llvmlite==0.42.0
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.8.4
mccabe==0.6.1
mdurl==0.1.1
mpmath==1.3.0
multianndata==0.0.4
natsort==8.4.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pbr==5.9.0
pillow==10.3.0
platformdirs==2.5.2
psutil==5.9.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pynndescent==0.5.12
pyparsing==3.1.2
pyreadr==0.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.27.1
scanpy==1.10.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
session_info==1.0.0
shap==0.45.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
statsmodels==0.14.2
stdlib-list==0.10.0
stevedore==3.5.0
sympy==1.12
threadpoolctl==3.5.0
toml==0.10.2
torch==2.3.0
tornado==6.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==1.26.9
xgboost==2.0.3
